requirements.txt
django
djangorestframework
django-cors-headers
-q -U google-genai
selenium
webdriver-manager
rapidfuzz
